requirements.txt
farm-haystack[inference,weaviate]==1.21.0
streamlit
notion-haystack
python-dotenv
